Refactor Floating Action Button Functionality
