+++ app/src/androidTest/java/com/github/mateo762/myapplication/CreateHabitActivityTest.kt
@@ -2,13 +2,19 @@ package com.github.mateo762.myapplication
 
 import androidx.compose.ui.test.*
 import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.fragment.app.testing.FragmentScenario
+import androidx.fragment.app.testing.launchFragment
 import androidx.lifecycle.Lifecycle
+import androidx.test.core.app.ActivityScenario
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.Intents.intended
 import androidx.test.espresso.intent.matcher.IntentMatchers.*
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.mateo762.myapplication.habits.HabitsActivity
+import com.github.mateo762.myapplication.home.HomeActivity
+import com.github.mateo762.myapplication.home.fragments.SummaryFragment
 import org.hamcrest.Matchers.*
 import org.junit.After
 import org.junit.Before
@@ -23,7 +29,9 @@ import java.util.*
 class CreateHabitActivityTest {
     /*
     @get:Rule
-    val activityRule = ActivityScenarioRule(CreateHabitActivity::class.java)
+    val activityRule = ActivityScenarioRule(HomeActivity::class.java)
+    val s = ActivityScenario(HomeActivity(), SummaryFragment())
+    val fs = FragmentScenario<SummaryFragment>(SummaryFragment(), HomeActivity())
 
     @get:Rule
     val composeTestRule = createComposeRule()
@@ -158,8 +166,7 @@ class CreateHabitActivityTest {
         composeTestRule.onNodeWithTag("btn_cancel")
             .performScrollTo()
             .performClick()
-        intended(hasComponent(MainActivity::class.java.name))
+        intended(hasComponent(HabitsActivity::class.java.name))
     }
-
-     */
+*/
 }
