+++ app/src/main/res/layout/fragment_list.xml
@@ -4,4 +4,11 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".habits.fragments.ListFragment">
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:gravity="center"
+        android:text="List Fragment"
+        android:textSize="20dp"/>
 </FrameLayout>
